refactor api mapping and fetching functions
