+++ frontend-next-migration/src/widgets/SectionMembers/model/membersApi.ts
@@ -1,54 +1,8 @@
+// membersApi.ts
+
 import { envHelper } from '@/shared/const/envHelper';
-import { Member, Department, Team } from '../model/types';
-
-// Function to map members
-const mapMembers = (membersData: any[]): Member[] => {
-  return (
-    membersData.map((member: any) => ({
-      id: member.id,
-      name: member.attributes.Name,
-      task: member.attributes.Task,
-      email: member.attributes.Email,
-      linkedin: member.attributes.Linkedin,
-      website: member.attributes.Website,
-      github: member.attributes.Github,
-      logo: member.attributes.Logo,
-      facebook: member.attributes.Facebook,
-      instagram: member.attributes.Instagram,
-      createdAt: member.attributes.createdAt,
-      updatedAt: member.attributes.updatedAt,
-      locale: member.attributes.locale,
-    })) || []
-  );
-};
-
-// Function to map departments and ensure the correct members are assigned
-const mapDepartments = (
-  departmentsData: any[],
-  locale: string,
-): Department[] => {
-  return (
-    departmentsData.map((dept: any) => {
-      // Find the localized name if available
-      const localizedDept = dept.attributes.localizations?.data.find(
-        (loc: any) => loc.attributes.locale === locale,
-      );
-
-      const localizedDeptName = localizedDept
-        ? localizedDept.attributes.Name
-        : dept.attributes.Name;
-
-      // Map the members that are explicitly part of this department
-      const members = mapMembers(dept.attributes.members?.data || []);
-
-      return {
-        id: dept.id,
-        name: localizedDeptName,
-        members,
-      };
-    }) || []
-  );
-};
+import { Member, Team } from '../model/types';
+import { mapMembers, mapDepartments } from '../model/mappers'; // Import mappers
 
 // Fetch Teams
 export const fetchTeams = async (locale: string = 'en'): Promise<Team[]> => {
